fix `VOID_TEXT` extractor & refactor `*Extractor` classes: * new `*ExtractorMixin` classes * new `*VoidExtractor` classes * move `Extractor` - `GeneratorCSSVoidExtractor` classes to `base_extractor` module
